fix: avoid circular import
Moving this enum to utils to avoid circular import error
